requirements.txt
Django	==2.2.2
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.1.0
djangorestframework	==3.9.4
idna	==2.8
requests==2.22.0
setuptools==39.1.0
urllib3==1.25.3
psycopg2 ==2.8.3
gunicorn==19.9.0
python-memcached==1.59